refactor: remove CruiseSider
