+++ src/components/CruiseMain/components/CruiseMenu/index.tsx
@@ -1,10 +1,21 @@
 import React, { useContext, useState } from 'react';
 import { BarsOutlined } from '@ant-design/icons';
 import styled from 'styled-components';
-import { Button, Drawer } from 'antd';
-import CruiseSider from './components/CruiseSider';
+import { Button, Drawer, Layout } from 'antd';
 import MenuContent from './components/MenuContent';
 import DeviceContext from '../../../../store/DeviceContext';
+
+const { Sider } = Layout;
+
+const StyledSider = styled(Sider)`
+  .ant-layout-sider-zero-width-trigger {
+    position: absolute;
+    top: -50px;
+    left: 0;
+    background: transparent;
+    color: black;
+  }
+`;
 
 const StyledButton = styled(Button)`
   font-size: 20px;
@@ -40,19 +51,29 @@ const CruiseMenu = (): JSX.Element => {
     setOpen(false);
   };
 
-  return deviceType === 'desktop' ? <CruiseSider /> : (
-    <>
-      <StyledButton type="link" icon={<BarsOutlined />} onClick={showDrawer} />
-      <StyledDrawer
-        placement="left"
-        closable={false}
-        onClose={onClose}
-        open={open}
+  return deviceType === 'desktop'
+    ? (
+      <StyledSider
+        theme="dark"
+        width={250}
+        collapsible={false}
       >
         <MenuContent />
-      </StyledDrawer>
-    </>
-  );
+      </StyledSider>
+    )
+    : (
+      <>
+        <StyledButton type="link" icon={<BarsOutlined />} onClick={showDrawer} />
+        <StyledDrawer
+          placement="left"
+          closable={false}
+          onClose={onClose}
+          open={open}
+        >
+          <MenuContent />
+        </StyledDrawer>
+      </>
+    );
 };
 
 export default CruiseMenu;
